Add new utility class to handle saved filter
